mcp_ffmpeg/pyproject.toml
[project]
name = "mcp-ffmpeg"
version = "0.1.0"
description = "An MCP server for FFmpeg operations with natural language interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [ "mcp>=1.1.1", "anthropic>=0.7.0", "python-dotenv>=1.0.0"]
[[project.authors]]
name = "chaorders1"
email = "1031luy@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-ffmpeg = "mcp_ffmpeg:server_main"
mcp-ffmpeg-client = "mcp_ffmpeg:client_main"
